requirements.txt
alembic==1.0.7
Click==7.0
Flask==1.0.2
Flask-Migrate==2.3.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
gunicorn==19.9.0
itsdangerous==1.1.0
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.1.0
psycopg2==2.7.7
python-dateutil==2.7.5
python-dotenv==0.10.1
python-editor==1.0.3
six==1.12.0
SQLAlchemy==1.2.17
Werkzeug==0.14.1
